requirements.txt
# AI模型依赖
openai>=1.0.0          # 用于 Deepseek API
google-generativeai>=0.3.0  # 用于 Gemini API
openai-whisper>=20231117    # 用于语音识别
sounddevice>=0.4.6     # 用于录音
soundfile>=0.12.1      # 用于音频文件处理
numpy>=1.24.0          # 用于音频处理
edge-tts>=6.1.9        # 用于语音合成
tqdm>=4.65.0           # 用于进度显示
rich>=13.7.0           # 用于彩色终端输出

# 工具依赖
python-dotenv>=1.0.0   # 用于环境变量管理
python-json-logger>=2.0.0  # 用于JSON格式日志
mysql-connector-python>=8.0.0  # MySQL数据库连接

# 日志相关
python-json-logger>=2.0.0  # 可选，用于JSON格式日志 